More precision for duration and latency

+++ src/Hangfire.Core/Storage/Monitoring/SucceededJobDto.cs
@@ -28,7 +28,7 @@ namespace Hangfire.Storage.Monitoring
 
         public Job Job { get; set; }
         public object Result { get; set; }
-        public long? TotalDuration { get; set; }
+        public double? TotalDuration { get; set; }
         public DateTime? SucceededAt { get; set; }
         public bool InSucceededState { get; set; }
     }
